refactor: 파일 구조 변경, user Detail 라우팅 추가
