+++ src/pages/UserManage/User/UserDetail.style.tsx
