+++ src/pages/UserManage/User/UserDetail.tsx
@@ -1,0 +1,16 @@
+import { useParams } from 'react-router-dom';
+import * as S from './UserDetail.style';
+
+function UserDetail() {
+  const { id } = useParams();
+  return (
+    <S.Container>
+      <S.Heading>
+        User:
+        {id}
+      </S.Heading>
+    </S.Container>
+  );
+}
+
+export default UserDetail;
