+++ src/pages/UserManage/User/UserList.style.tsx
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const Container = styled.div`
+  
+`;
+
+export const Heading = styled.h1`
+  
+`;
